requirements.txt
weaviate-client==4.4b0
pandas==1.5.0